Split query header out to separate component
