+++ src/components/Query.js
@@ -2,51 +2,12 @@ import qs from 'query-string';
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import DropSelect from './DropSelect';
 import Loading from './Loading';
+import QueryHeader from './QueryHeader';
 import Tag from './Tag';
 import TicketList from './TicketList';
-import LabelSelect from '../containers/selectors/LabelSelect';
-import MilestoneSelect from '../containers/selectors/MilestoneSelect';
 
 import './Query.css';
-
-const SORT_OPTIONS = [
-	{
-		id: 'new',
-		title: 'Newest',
-		value: {
-			order: 'time',
-			desc: '1',
-		},
-	},
-	{
-		id: 'old',
-		title: 'Oldest',
-		value: {
-			order: 'time',
-			desc: '0',
-		},
-	},
-	{
-		id: 'new-update',
-		title: 'Most recently updated',
-		value: {
-			order: 'changetime',
-			desc: '1',
-		},
-	},
-	{
-		id: 'old-update',
-		title: 'Least recently updated',
-		value: {
-			order: 'changetime',
-			desc: '0',
-		},
-	},
-];
-
-const Label = ({ text }) => <span>{ text } <span className="Query-drop-arrow">▼</span></span>;
 
 export default class Query extends React.PureComponent {
 	constructor( props ) {
@@ -80,47 +41,13 @@ export default class Query extends React.PureComponent {
 
 		const page = params.page ? parseInt( params.page, 10 ) : 1;
 
-		const currentOrder = params.order || 'time';
-		const currentOrderDesc = params.desc || '1';
-
-		const matchedSort = SORT_OPTIONS.find( opt => {
-			return currentOrder === opt.value.order && currentOrderDesc === opt.value.desc;
-		});
-		const currentSort = matchedSort ? matchedSort.id : '';
-
 		return <div className="Query">
 			<h1>Query</h1>
-			<div className="Query-header">
-				<div className="Query-count" />
-				<nav>
-					<ul className="Query-filter">
-						<li>
-							<LabelSelect
-								label={ <Label text="Labels" /> }
-								selected={ params.keywords }
-								onSelect={ value => onUpdateQuery( value ) }
-							/>
-						</li>
+			<QueryHeader
+				params={ params }
+				onUpdateQuery={ onUpdateQuery }
+			/>
 
-						<li>
-							<MilestoneSelect
-								label={ <Label text="Milestones" /> }
-								selected={ params.milestone }
-								onSelect={ milestone => onUpdateQuery({ milestone }) }
-							/>
-						</li>
-
-						<li>
-							<DropSelect
-								label={ <Label text="Sort" /> }
-								items={ SORT_OPTIONS }
-								selected={ currentSort }
-								onSelect={ value => onUpdateQuery( value ) }
-							/>
-						</li>
-					</ul>
-				</nav>
-			</div>
 			{ loading ? (
 				<Loading />
 			) : (
